requirements.txt
torch
torchvision
pandas
seaborn
matplotlib
scikit-learn
Pillow
ultralytics